ui: move timeline view to its own file
Also add some thread methods for fetching new notes
